+++ source/parsing/SyntaxNode.cpp
@@ -47,8 +47,8 @@ Token SyntaxNode::getFirstToken() const {
 
 Token SyntaxNode::getLastToken() const {
     uint32_t childCount = getChildCount();
-    for (int i = childCount - 1; i >= 0; i--) {
-        auto child = getChild(i);
+    for (int i = int(childCount - 1); i >= 0; i--) {
+        auto child = getChild(uint32_t(i));
         if (child.isToken) {
             if (child.token)
                 return child.token;
